backend/requirement.txt
flask==2.2.5
flask-cors==3.0.10
pdf2image==1.16.3
pillow==9.4.0
google-generativeai
python-dotenv
pdf2image